:mobile_phone_off: add feature toggle to be able to disable delete action

+++ Feedback/NHS111.Business.Feedback.Api/Controllers/FeedbackController.cs
@@ -2,6 +2,7 @@
 using System.Net.Http;
 using System.Threading.Tasks;
 using System.Web.Http;
+using NHS111.Business.Feedback.Api.Features;
 using NHS111.Domain.Feedback.Models;
 using NHS111.Domain.Feedback.Repository;
 using NHS111.Utils.Attributes;
@@ -32,23 +33,34 @@ namespace NHS111.Business.Feedback.Api.Controllers
         [Route("delete/{identifier}")]
         public async Task<HttpResponseMessage> DeleteFeedback(string identifier)
         {
-            var result = (DatabaseCode) await _feedbackRepository.Delete(identifier);
+            DeleteFeedbackFeature featureToggle = new DeleteFeedbackFeature();
             HttpResponseMessage response;
 
-            if (result == DatabaseCode.Success)
+            if (featureToggle.IsEnabled)
             {
-                var responseMessage = string.Format("Row with id {0} was successfully deleted.", identifier);
-                response = Request.CreateResponse(System.Net.HttpStatusCode.OK, responseMessage);
+                var result = (DatabaseCode)await _feedbackRepository.Delete(identifier);
+
+                if (result == DatabaseCode.Success)
+                {
+                    var responseMessage = string.Format("Row with id {0} was successfully deleted.", identifier);
+                    response = Request.CreateResponse(System.Net.HttpStatusCode.OK, responseMessage);
+                }
+                else
+                {
+                    var responseMessage = string.Format("Row with id {0} could not be found.", identifier);
+                    response = Request.CreateResponse(System.Net.HttpStatusCode.NotFound, responseMessage);
+                }
+
+                return response;
             }
             else
             {
-                var responseMessage = string.Format("Row with id {0} could not be found.", identifier);
-                response = Request.CreateResponse(System.Net.HttpStatusCode.NotFound, responseMessage);
+                response = Request.CreateResponse(System.Net.HttpStatusCode.NotFound);
+                return response;
             }
-            
-            return response;
+
         }
-        
+
         [HttpGet]
         [Route("list")]
         public async Task<IEnumerable<Domain.Feedback.Models.Feedback>> ListFeedback()
